# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "AI Chatbot",
        "version": "0.1.0"
    },
    "paths": {
        "/v1/chat/completions": {
            "post": {
                "tags": [
                    "chat"
                ],
                "summary": "Create Chat Completion",
                "operationId": "createChatCompletion",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChatCompletionRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/chat/{chat_id}": {
            "get": {
                "tags": [
                    "chat"
                ],
                "summary": "Get Chat",
                "operationId": "getChat",
                "parameters": [
                    {
                        "name": "chat_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Chat Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Chat"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "chat"
                ],
                "summary": "Delete Chat",
                "operationId": "deleteChat",
                "parameters": [
                    {
                        "name": "chat_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Chat Id"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful Response"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/chat/": {
            "get": {
                "tags": [
                    "chat"
                ],
                "summary": "Get Chats",
                "operationId": "getChats",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/Chat"
                                    },
                                    "type": "array",
                                    "title": "Response Getchats"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/chat/{chat_id}/messages": {
            "post": {
                "tags": [
                    "chat"
                ],
                "summary": "Create Message",
                "operationId": "createMessage",
                "parameters": [
                    {
                        "name": "chat_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Chat Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Message"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "tags": [
                    "chat"
                ],
                "summary": "Get Messages",
                "operationId": "getMessages",
                "parameters": [
                    {
                        "name": "chat_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Chat Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Message"
                                    },
                                    "title": "Response Getmessages"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/files/upload": {
            "post": {
                "tags": [
                    "files"
                ],
                "summary": "Upload File",
                "operationId": "uploadFile",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/Body_uploadFile"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FileResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/": {
            "get": {
                "summary": "Root",
                "operationId": "root__get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "title": "Response Root  Get"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Body_uploadFile": {
                "properties": {
                    "chat_id": {
                        "type": "string",
                        "title": "Chat Id",
                        "description": "The chatId the uploaded file."
                    },
                    "file": {
                        "type": "string",
                        "format": "binary",
                        "title": "File",
                        "description": "The file to be uploaded."
                    }
                },
                "type": "object",
                "required": [
                    "chat_id",
                    "file"
                ],
                "title": "Body_uploadFile"
            },
            "Chat": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Id",
                        "description": "The unique identifier of the chat."
                    },
                    "title": {
                        "type": "string",
                        "title": "Title",
                        "description": "The title of the chat."
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "title"
                ],
                "title": "Chat"
            },
            "ChatCompletionRequest": {
                "properties": {
                    "model": {
                        "type": "string",
                        "title": "Model",
                        "description": "The model to use for generating completions (e.g., 'gpt-3.5-turbo')."
                    },
                    "messages": {
                        "items": {
                            "$ref": "#/components/schemas/Message"
                        },
                        "type": "array",
                        "title": "Messages",
                        "description": "A list of messages describing the conversation history."
                    },
                    "max_tokens": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Max Tokens",
                        "description": "The maximum number of tokens to generate in the response."
                    },
                    "temperature": {
                        "anyOf": [
                            {
                                "type": "number",
                                "maximum": 2,
                                "minimum": 0
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Temperature",
                        "description": "Sampling temperature. Higher values make output more random; lower values make it more deterministic.",
                        "default": 1
                    },
                    "top_p": {
                        "anyOf": [
                            {
                                "type": "number",
                                "maximum": 1,
                                "minimum": 0
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Top P",
                        "description": "Nucleus sampling probability. The model considers results of tokens with top_p probability mass.",
                        "default": 1
                    },
                    "stream": {
                        "anyOf": [
                            {
                                "type": "boolean"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Stream",
                        "description": "Whether to return a stream of data for the response. Defaults to False.",
                        "default": false
                    },
                    "stop": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "items": {
                                    "type": "string"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Stop",
                        "description": "Optional stop sequences where the model will stop generating further tokens."
                    },
                    "presence_penalty": {
                        "anyOf": [
                            {
                                "type": "number",
                                "maximum": 2,
                                "minimum": -2
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Presence Penalty",
                        "description": "Penalty for new tokens based on whether they appear in the text so far.",
                        "default": 0
                    },
                    "frequency_penalty": {
                        "anyOf": [
                            {
                                "type": "number",
                                "maximum": 2,
                                "minimum": -2
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Frequency Penalty",
                        "description": "Penalty for new tokens based on their frequency in the text so far.",
                        "default": 0
                    },
                    "user": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "User",
                        "description": "Unique identifier for the end-user. This helps OpenAI monitor and detect abuse."
                    }
                },
                "type": "object",
                "required": [
                    "model",
                    "messages"
                ],
                "title": "ChatCompletionRequest"
            },
            "FileResponse": {
                "properties": {
                    "id": {
                        "type": "string",
                        "title": "Id"
                    },
                    "object": {
                        "type": "string",
                        "title": "Object"
                    },
                    "bytes": {
                        "type": "integer",
                        "title": "Bytes"
                    },
                    "created_at": {
                        "type": "integer",
                        "title": "Created At"
                    },
                    "filename": {
                        "type": "string",
                        "title": "Filename"
                    },
                    "purpose": {
                        "type": "string",
                        "title": "Purpose"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "object",
                    "bytes",
                    "created_at",
                    "filename",
                    "purpose"
                ],
                "title": "FileResponse"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "Message": {
                "properties": {
                    "id": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "uuid"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Id"
                    },
                    "role": {
                        "type": "string",
                        "title": "Role",
                        "description": "The role of the message sender (e.g., 'user', 'assistant', or 'system')."
                    },
                    "content": {
                        "type": "string",
                        "title": "Content",
                        "description": "The content of the message."
                    }
                },
                "type": "object",
                "required": [
                    "role",
                    "content"
                ],
                "title": "Message"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}